test(unit tests): restructure tests to optimize parallelization
